+++ R-package/src/rust/api.h
@@ -27,6 +27,8 @@ SEXP rep_str_vec(SEXP x);
 SEXP rep_str_slice(SEXP x);
 SEXP safe_stop(void);
 SEXP raise_error(void);
+SEXP do_call(SEXP fun, SEXP args, SEXP env);
+SEXP call_with_args(SEXP fun, SEXP env);
 SEXP new_int(SEXP size);
 SEXP new_real(SEXP size);
 SEXP new_bool(SEXP size);