+++ src/lib.rs
@@ -13,6 +13,8 @@ pub mod unwind_protect;
 use std::os::raw::c_char;
 
 pub use error::{Error, Result};
+pub use sexp::environment::EnvironmentSexp;
+pub use sexp::external_pointer::{get_external_pointer_addr, ExternalPointerSexp, IntoExtPtrSexp};
 pub use sexp::integer::{IntegerSexp, OwnedIntegerSexp};
 pub use sexp::list::{ListSexp, OwnedListSexp};
 pub use sexp::logical::{LogicalSexp, OwnedLogicalSexp};
@@ -20,8 +22,6 @@ pub use sexp::null::NullSexp;
 pub use sexp::real::{OwnedRealSexp, RealSexp};
 pub use sexp::string::{OwnedStringSexp, StringSexp};
 pub use sexp::{Sexp, TypedSexp};
-
-pub use sexp::external_pointer::{get_external_pointer_addr, ExternalPointerSexp, IntoExtPtrSexp};
 
 pub use unwind_protect::unwind_protect;
 
@@ -31,8 +31,8 @@ pub use savvy_macro::savvy;
 use ffi::SEXP;
 use savvy_ffi::{cetype_t_CE_UTF8, Rf_allocVector, Rf_mkCharLenCE};
 
-fn alloc_vector(arg1: u32, arg2: isize) -> crate::error::Result<SEXP> {
-    unsafe { unwind_protect(|| Rf_allocVector(arg1, arg2)) }
+fn alloc_vector(arg1: u32, arg2: usize) -> crate::error::Result<SEXP> {
+    unsafe { unwind_protect(|| Rf_allocVector(arg1, arg2 as _)) }
 }
 
 // This wrapper function handles Error and panicks, and flag it by setting the