+++ src/sexp/mod.rs
@@ -1,8 +1,8 @@
 use std::ffi::{CStr, CString};
 
 use savvy_ffi::{
-    R_NilValue, Rf_isInteger, Rf_isLogical, Rf_isReal, Rf_isString, Rf_type2char, EXTPTRSXP, SEXP,
-    TYPEOF, VECSXP,
+    R_NilValue, Rf_isEnvironment, Rf_isFunction, Rf_isInteger, Rf_isLogical, Rf_isReal,
+    Rf_isString, Rf_type2char, EXTPTRSXP, SEXP, TYPEOF, VECSXP,
 };
 
 use crate::{
@@ -10,7 +10,9 @@ use crate::{
     OwnedLogicalSexp, OwnedRealSexp, OwnedStringSexp, RealSexp, StringSexp,
 };
 
+pub mod environment;
 pub mod external_pointer;
+pub mod function;
 pub mod integer;
 pub mod list;
 pub mod logical;
@@ -64,6 +66,16 @@ impl Sexp {
         unsafe { TYPEOF(self.0) as u32 == EXTPTRSXP }
     }
 
+    /// Returns `true` if the SEXP is an environment.
+    pub fn is_environment(&self) -> bool {
+        unsafe { Rf_isEnvironment(self.0) == 1 }
+    }
+
+    /// Returns `true` if the SEXP is a function.
+    pub fn is_function(&self) -> bool {
+        unsafe { Rf_isFunction(self.0) == 1 }
+    }
+
     /// Returns the string representation of the SEXP type.
     #[allow(clippy::not_unsafe_ptr_arg_deref)]
     pub fn get_human_readable_type_name(&self) -> &'static str {
@@ -84,6 +96,8 @@ pub enum TypedSexp {
     List(ListSexp),
     Null(NullSexp),
     ExternalPointer(ExternalPointerSexp),
+    Environment(EnvironmentSexp),
+    Function(FunctionSexp),
     Other(SEXP),
 }
 
@@ -103,6 +117,8 @@ into_typed_sxp!(StringSexp, String);
 into_typed_sxp!(LogicalSexp, Logical);
 into_typed_sxp!(ListSexp, List);
 into_typed_sxp!(ExternalPointerSexp, ExternalPointer);
+into_typed_sxp!(EnvironmentSexp, Environment);
+into_typed_sxp!(FunctionSexp, Function);
 into_typed_sxp!(NullSexp, Null);
 
 macro_rules! into_typed_sxp_owned {
@@ -130,6 +146,8 @@ impl From<TypedSexp> for SEXP {
             TypedSexp::Logical(sxp) => sxp.inner(),
             TypedSexp::List(sxp) => sxp.inner(),
             TypedSexp::ExternalPointer(sxp) => sxp.inner(),
+            TypedSexp::Environment(sxp) => sxp.inner(),
+            TypedSexp::Function(sxp) => sxp.inner(),
             TypedSexp::Other(sxp) => sxp,
         }
     }
@@ -146,6 +164,11 @@ impl Sexp {
             savvy_ffi::LGLSXP => TypedSexp::Logical(LogicalSexp(self.0)),
             savvy_ffi::VECSXP => TypedSexp::List(ListSexp(self.0)),
             savvy_ffi::EXTPTRSXP => TypedSexp::ExternalPointer(ExternalPointerSexp(self.0)),
+            savvy_ffi::ENVSXP => TypedSexp::Environment(EnvironmentSexp(self.0)),
+            // cf. https://github.com/wch/r-source/blob/95ac44a87065d5b42579b621d278adc44641dcf0/src/include/Rinlinedfuns.h#L810-L815
+            savvy_ffi::CLOSXP | savvy_ffi::BUILTINSXP | savvy_ffi::SPECIALSXP => {
+                TypedSexp::Function(FunctionSexp(self.0))
+            }
             savvy_ffi::NILSXP => TypedSexp::Null(NullSexp),
             _ => TypedSexp::Other(self.0),
         }
@@ -367,3 +390,5 @@ macro_rules! impl_common_sexp_ops_owned {
 
 pub(crate) use impl_common_sexp_ops;
 pub(crate) use impl_common_sexp_ops_owned;
+
+use self::{environment::EnvironmentSexp, function::FunctionSexp};